requirements.txt
streamlit
requests
google-generativeai
python-dotenv
pytz
sentence_transformers
faiss-cpu